refactor: (#111) queryFactory user -> student
