+++ dms-application/src/main/kotlin/team/aliens/dms/domain/manager/spi/ManagerQueryStudentPort.kt
@@ -1,10 +1,15 @@
 package team.aliens.dms.domain.manager.spi
 
+import team.aliens.dms.domain.manager.dto.Sort
 import team.aliens.dms.domain.student.model.Student
 import java.util.UUID
 
 interface ManagerQueryStudentPort {
 
     fun queryStudentById(studentId: UUID): Student?
+
+    fun queryUserByNameAndSort(name: String, sort: Sort): List<Student>
+
+    fun queryUserByRoomNumberAndSchoolId(roomNumber: Int, schoolId: UUID): List<Student>
 
 }
