+++ dms-application/src/main/kotlin/team/aliens/dms/domain/manager/spi/ManagerQueryUserPort.kt
@@ -1,6 +1,5 @@
 package team.aliens.dms.domain.manager.spi
 
-import team.aliens.dms.domain.manager.dto.Sort
 import team.aliens.dms.domain.user.model.User
 import java.util.UUID
 
@@ -13,9 +12,5 @@ interface ManagerQueryUserPort {
     fun queryUserByAccountId(accountId: String): User?
 
     fun queryUserByEmail(email: String): User?
-
-    fun queryUserByNameAndSort(name: String, sort: Sort): List<User>
-    
-    fun queryUserByRoomIdAndSchoolId(roomId: UUID, schoolId: UUID): List<User>
 
 }
