+++ dms-infrastructure/src/main/kotlin/team/aliens/dms/persistence/user/UserPersistenceAdapter.kt
@@ -1,24 +1,17 @@
 package team.aliens.dms.persistence.user
 
-import com.querydsl.core.types.OrderSpecifier
-import com.querydsl.jpa.impl.JPAQueryFactory
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Component
-import team.aliens.dms.domain.manager.dto.Sort
 import team.aliens.dms.domain.user.model.User
 import team.aliens.dms.domain.user.spi.UserPort
 import team.aliens.dms.persistence.user.mapper.UserMapper
 import team.aliens.dms.persistence.user.repository.UserJpaRepository
 import java.util.UUID
 
-import team.aliens.dms.persistence.student.entity.QStudentJpaEntity.studentJpaEntity
-import team.aliens.dms.persistence.user.entity.QUserJpaEntity.userJpaEntity
-
 @Component
 class UserPersistenceAdapter(
     private val userMapper: UserMapper,
-    private val userRepository: UserJpaRepository,
-    private val queryFactory: JPAQueryFactory
+    private val userRepository: UserJpaRepository
 ) : UserPort {
 
     override fun existsUserByEmail(email: String) = userRepository.existsByEmail(email)
@@ -46,47 +39,5 @@ class UserPersistenceAdapter(
     override fun queryUserByAccountId(accountId: String) = userMapper.toDomain(
         userRepository.findByAccountId(accountId)
     )
-
-    override fun queryUserByNameAndSort(name: String, sort: Sort): List<User> {
-        return queryFactory
-            .selectFrom(userJpaEntity)
-            .join(studentJpaEntity).on(userJpaEntity.id.eq(studentJpaEntity.userId))
-            .where(
-                userJpaEntity.name.like(name)
-            )
-            .orderBy(
-                sortFilter(sort)
-            )
-            .fetch()
-            .map {
-                userMapper.toDomain(it)!!
-            }
-    }
-
-    override fun queryUserByRoomIdAndSchoolId(roomId: UUID, schoolId: UUID): List<User> {
-        return queryFactory
-            .selectFrom(userJpaEntity)
-            .join(studentJpaEntity).on(userJpaEntity.id.eq(studentJpaEntity.userId))
-            .where(
-                studentJpaEntity.,
-                studentJpaEntity.room.id.schoolId.eq(schoolId)
-            ).fetch()
-            .map {
-                userMapper.toDomain(it)!!
-            }
-    }
-
-    private fun sortFilter(sort: Sort): OrderSpecifier<*>? {
-        return when(sort) {
-            Sort.NAME -> {
-                userJpaEntity.name.asc()
-            }
-            else -> {
-                studentJpaEntity.grade.asc()
-                studentJpaEntity.classRoom.asc()
-                studentJpaEntity.number.asc()
-            }
-        }
-    }
 }
